Add handling for invalid commands in main.cc (#750)
* Add handling for invalid commands in main.cc

Added a default case to the switch statement in the main function to
handle invalid commands.
This improves user experience by providing guidance when an invalid
command is used.

* Fix Protobuf library shutdown placement

The protobuf shutdown was previously never called, given that all switch
cases returned before it.
This commit introduces a new variable to store the return value to store
the command execution, and returns after the proper shutdown of the Protobuf
library to prevent resource leaks