translation: label REGIONAL_FAST without regional (#1067)
as the term regional may confuse one into thinking the train can be used
with a ticket of regional public transport. In Germany there was special
confusion, as for those routes I checked, the class of RegionalExpress
trains in the data is labeled with REGIONAL_RAIL (IMHO correctly), and
REGIONAL_FAST_RAIL was translated with RegionalExpress. I avoided to use
the term InterRegio, as in Switzerland those are REGIONAL_RAIL, but the
German railway company uses the term "InterRegio / fast train" for the
lowest class of non-regional long distance. AFAIK German railway uses
the "fast train" classification only for foreign trains.

The definition in openapi.yaml matches with
https://en.wikipedia.org/wiki/Express_train and a few of the other
languages linked for that article. And I have not seen the term further
differentiated with something like regional, only with a specific line.

I'm told REGIONAL_FAST_RAIL is also used as a fallback where there is no data and
that the structure of the data is
not expressive enough to capture the ticket price structure, but as it
is a few test routings on https://transitous.org/ revealed no
conflicting usage of REGIONAL_FAST_RAIL. (Whether or not some of these
should be a "faster" classification, does not pose any problem here.)

Whether it is because the origin on the data already fits this
classification system and/or the fall back is selected well, but overall
I think omitting regional is more likely to lead people to select the
route options they intent.